backend/requirements.txt
# Framework web
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de datos
sqlalchemy==2.0.35
pymysql==1.1.1

# Seguridad y autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # ✅ CAMBIAR DE 4.2.0 a 4.1.2
cryptography==43.0.1

# Variables de entorno
python-dotenv==1.0.1

# Cliente HTTP
httpx==0.27.2

# OpenAI (opcional)
openai==1.51.2

# Validación de datos
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilidades
python-dateutil==2.9.0

# Ngrok
pyngrok==7.2.0